Move reference classification to ra_ide_db
Lost some marks along the way :-(
